début création des fichier Sprite.h
