+++ src/AllSprite.h
@@ -1,0 +1,14 @@
+//
+//  AllSprite.h
+//  ProjetMapping
+//
+//  Created by FatDazz_mac on 30/03/2018.
+//
+//
+
+#pragma once
+
+#include "SpriteAvatar.h"
+#include "SpriteBlock.h"
+#include "SpritePickup.h"
+#include "Sprite.h"
