+++ src/Box2D/PhysicalizedElement/ObjSignageSign.cpp
@@ -1,0 +1,9 @@
+//
+//  ObjSignageSign.cpp
+//  ProjetMapping
+//
+//  Created by FatDazz_mac on 29/03/2018.
+//
+//
+
+#include "ObjSignageSign.h"
