+++ src/Box2D/PhysicalizedElement/Teleportable/ObjPlatforMove.cpp
@@ -1,0 +1,9 @@
+//
+//  ObjPlatforMove.cpp
+//  ProjetMapping
+//
+//  Created by FatDazz_mac on 29/03/2018.
+//
+//
+
+#include "ObjPlatforMove.h"
