+++ src/Sprite.h
@@ -82,95 +82,6 @@ public:
     
 };
 
-class Avatar;
-class AvatarDef: public SpriteObj, public Light{
-public:
-    AvatarDef(){
-        categoryBits = Category::AVATAR;
-        maskBits  =  Category::AVATAR  | Category::PORTAL | Category::LADDER |Category::OBJ | Category::CLONE;
-        layerIni = layerId =  3;
-        color.set(ofColor::orange);
-        drawMove = &AvatarDef::drawMovePlaform;
-        moveImages[0].load("ava_1.png");
-        //moveImages[0].mirror(false, false);
-        moveImages[1].load("ava_2.png");
-        moveImages[2].load("ava_3.png");
-        moveImages[3].load("ava_4.png");
-        
-    }
-    ofImage spriteImage;
-    Shift* s;
-    Avatar* a;
-    void draw()override;
-    ofVec2f getPositionTranform() override;
-    void create() override;
-    void reset() override;
-    bool isActif() override;
-    bool getViewPoint() override;
-    
-    
-    void setMove(int i);//Deplacement move);
-    void drawMovePlaform(ofVec2f p);
-    void drawMoveTop(ofVec2f p);
-
-    void(AvatarDef::*drawMove)(ofVec2f p);
-
-    
-    void drawjump();
-    
-    float n;
-    ofImage moveImages[4];
-    ofImage jumpImage[3];
-    ofImage jadderImage[3];
-
-    
-};
-
-class ObjBlock;
-class ObjBlockDef: public SpriteObj{
-public:
-    ObjBlockDef(){
-        categoryBits = Category::OBJ;
-        maskBits = Category::PORTAL | Category::AVATAR | Category::OBJ | Category::CLONE ;
-        layerIni = layerId =  2;
-    }
-    ofImage spriteImage;
-    ObjBlock * b;
-    void draw() override;
-    ofVec2f getPositionTranform() override;
-    void create() override;
-    void reset() override;
-    bool isActif() override;
-    bool getViewPoint() override;
-
-    
-};
-
-class ObjPickup;
-class ObjPickupDef: public SpriteObj, public Light {
-public:
-    
-    ObjPickupDef(){
-        categoryBits = Category::OBJ;
-        maskBits = Category::PORTAL | Category::AVATAR | Category::OBJ  | Category::CLONE;
-        radius = 0.2;
-        layerIni = layerId =  3;
-        color.set(ofColor::yellow);
-    }
-    bool actif  = true;
-    ofImage spriteImage;
-    ObjPickup * pkup;
-    void draw() override;
-
-    
-
-    ofVec2f getPositionTranform() override;
-    void create() override;
-    void reset() override;
-    bool isActif() override;
-    bool getViewPoint() override;
-
-};
 
 class ObjMushroom;
 class ObjMushroomDef: public SpriteObj{
